PDFBOX-5660: DRY refactoring
git-svn-id: https://svn.apache.org/repos/asf/pdfbox/branches/3.0@1913198 13f79535-47bb-0310-9956-ffa450edef68
